refactor: TUI起動とクリーンアップの重複コードを共通化
DRY原則に従い、TUIアプリ起動ロジックをユーティリティ化:

重複していた処理:
- render()の実行とクリーンアップ処理
- SIGINT/SIGTERMハンドラーの設定
- エラーハンドリングとプロセス終了処理

解決策:
- src/apps/cli/utils/tui-launcher.ts: 共通のTUI起動ユーティリティを作成
- launchTuiApp(): 安全なTUI起動とクリーンアップを提供
- tutorial.ts, tui.tsx: 重複コードを削除して共通関数を使用

メリット:
- 保守性向上: 一箇所を修正すれば全体に反映
- バグリスク軽減: ロジックの不整合を防止
- テスタビリティ改善: 共通ロジックの単体テストが可能

修正後のコード量:
- tutorial.ts: 16行削減
- tui.tsx: 10行削減
- 新規ユーティリティ: 44行（net効果: -82行）
