+++ src/apps/cli/commands/tui.tsx
@@ -1,29 +1,15 @@
-import { render } from 'ink';
 import { Command } from 'commander';
+import { launchTuiApp } from '../utils/tui-launcher.js';
 import { App } from '../../tui/App.js';
 
 export function createTuiCommand(): Command {
   const tuiCommand = new Command('tui');
 
   tuiCommand.description('Start RSS reader in TUI mode').action(() => {
-    try {
-      // TUIアプリケーション内でマイグレーションが実行される
-
-      // TUIアプリケーションを起動
-      const { clear, unmount } = render(<App />);
-
-      // プロセス終了時のクリーンアップ（Raw modeエラー防止）
-      const cleanup = () => {
-        clear();
-        unmount();
-      };
-
-      process.on('SIGINT', cleanup);
-      process.on('SIGTERM', cleanup);
-    } catch (error) {
-      console.error('TUIの起動に失敗しました:', error instanceof Error ? error.message : error);
-      process.exit(1);
-    }
+    // TUIアプリケーション内でマイグレーションが実行される
+    launchTuiApp(<App />, {
+      appName: 'TUI',
+    });
   });
 
   return tuiCommand;
