+++ src/apps/cli/commands/tutorial.ts
@@ -1,7 +1,7 @@
 import { Command } from 'commander';
 import { createDatabaseManager } from '../utils/database.js';
 import { createFeedServices } from '../../../services/factory.js';
-import { render } from 'ink';
+import { launchTuiApp } from '../utils/tui-launcher.js';
 import React from 'react';
 import { App } from '../../tui/App.js';
 
@@ -37,27 +37,11 @@ export function createTutorialCommand(): Command {
       }
 
       // TUIモードを起動
-      try {
-        // @ts-expect-error - カスタムpropsを渡すため
-        const { clear, unmount } = render(React.createElement(App, { databaseManager }));
-
-        // プロセス終了時のクリーンアップ
-        const cleanup = () => {
-          clear();
-          unmount();
-          databaseManager.close();
-        };
-
-        process.on('SIGINT', cleanup);
-        process.on('SIGTERM', cleanup);
-      } catch (error) {
-        console.error(
-          'チュートリアルの起動に失敗しました:',
-          error instanceof Error ? error.message : error
-        );
-        databaseManager.close();
-        process.exit(1);
-      }
+      // @ts-expect-error - カスタムpropsを渡すため
+      launchTuiApp(React.createElement(App, { databaseManager }), {
+        appName: 'チュートリアル',
+        databaseManager,
+      });
     });
 
   return command;
